refactor: Update Charts panel
